requirements.txt
# HTTP & Networking
httpx[http2]==0.27.*
aiohttp==3.9.*

# HTML/JS parsing
beautifulsoup4==4.12.*
lxml==5.1.*

# AI Agent
anthropic==0.49.*
openai==1.60.*

# Data & Config
pyyaml==6.0.*
pydantic==2.10.*
pydantic-settings==2.7.*

# Database (findings storage)
sqlalchemy==2.0.*
aiosqlite==0.20.*

# Rate limiting
aiolimiter==1.1.*

# Reporting
jinja2==3.1.*

# CLI
click==8.1.*
rich==13.9.*

# Testing
pytest==8.3.*
pytest-asyncio==0.24.*
